Check file open modes (read/write) in RPCs
Only allow setting a mode (and related stuff) when the client has opened
the node for writing, and reading back the mode if opened for reading.
This ensures the client actually has the authority to perform these
operations -- the open mode is checked against the file access bits on
open, by the standard trivfs implementation.

Probably it's not strictly necessary to check anything but the actual
kgi_set_mode() RPC, as none of the others are really destructive, or
exposing any important information. (kgi_get_image_mode() only obtains
information about the mode requested by the same client, so it shouldn't
hurt.)