Implement framebuffer mapping over RPC
Implemented the io_map() call to map the framebuffer, and two new KGI-specific
RPCs.

kgi_get_fb_resource() retrieves information about the framebuffer resource,
which the client needs to map the framebuffer. The information provided is
suitable for (later) use in kgiGetResource() in libggi.

kgi_setup_mmap_fb() is for use in kgiSetupMmapFB() in libggi. It is actually
rather pointless, as long as we don't implement mapping of anything else than
the framebuffer -- but we don't want to diverge from the standard KGI ioctl
interface too much...

Mapping the framebuffer is a bit tricky, as device_map() always gives a memory
object without any restrictions. Not only is it impossible to pass that to the
client for security consideration; but also, lacking the ability to specify an
offset, it's impossible for the client to mmap() the framebuffer directly. Thus
we need to create a proxy memory object from the one provided by device_map,
and pass that to the client instead.

The existing partial implementation of proxy memory objects (available as a
patch in the Debian gnumach package) only implements access permissions
(read/write) though, but doesn't implement the offset or range restrictions.
Thus, additional kernel patches are necessary to make this actually work.

This commit also adds some code to the rpctest program, so it maps the
framebuffer using the new RPCs, and draws some crap into it. (Just like the
server itself does when not started as translator...)