Save/restore VGA memory on driver load/unload
Save the contents of the standard VGA video memory, so that the text
console is usable again after the driver exits.

(Except that the palette is fucked up... But that's another issue.)

We save the whole 256 KiB of VGA memory. Only a fraction of this is
really relevant for the text console -- but saving it all makes it
simpler.