Handle CheckMode() failures
So far failures from CheckMode() were ignored. There was no error code
returned to the client; and the status was changed as if the CheckMode()
succeeded. As a consequence, SetMode() was called on an invalid mode
structure, resulting in driver segfault.

Added another set of calls to the rpctest program as well, to check
correct behaviour when trying to set an invalid mode.